Add getLog controller and separate into logsController file
