+++ server/controllers/flagsController.js
@@ -1,29 +1,13 @@
 const pg = require("pg");
 const { validationResult } = require("express-validator");
 const PGTable = require("../db/PGTable");
-const { FLAG_TABLE_NAME, FLAG_EVENTS_TABLE_NAME } = require("../constants/db");
+const { FLAG_TABLE_NAME } = require("../constants/db");
 const { getNowString } = require("../utils");
 const { sendWebhook } = require("../lib/sendWebhook.js");
 const { getFlagsForWebhook } = require("../db/flags.js");
 
 const flagTable = new PGTable(FLAG_TABLE_NAME);
-const flagEventsTable = new PGTable(FLAG_EVENTS_TABLE_NAME);
 flagTable.init();
-flagEventsTable.init();
-
-const logEvent = async (req, res, next) => {
-  try {
-    const event = {
-      flag_id: req.params.id || req.flagId,
-      event_type: req.eventType,
-      timestamp: getNowString()
-    };
-    const savedEvent = await flagEventsTable.insertRow(event);
-    res.status(200);
-  } catch (err) {
-    console.log(err.message);
-  }
-};
 
 const createNewFlagObj = ({
   name,
@@ -49,7 +33,6 @@ const createNewFlagObj = ({
 const getAllFlags = async (req, res, next) => {
   try {
     let data;
-
     if (req.query.prov) {
       data = await getFlagsForWebhook();
     } else {
@@ -167,4 +150,3 @@ exports.deleteFlag = deleteFlag;
 exports.getFlag = getFlag;
 exports.setFlagsOnReq = setFlagsOnReq;
 exports.sendFlagsWebhook = sendFlagsWebhook;
-exports.logEvent = logEvent;
